% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustats-package.R
\docType{package}
\name{ustats-package}
\alias{ustats}
\alias{ustats-package}
\title{ustats: R Interface to Python Tools for Computing Higher-Order U-Statistics}
\description{
Provides an R interface to the Python package 'u-stats' \url{https://pypi.org/project/u-stats/} for efficient computation of higher-order U-statistics using Einstein summation notation, implementing the methods of Chen, Zhang, and Liu (2025) \doi{10.48550/arXiv.2508.12627}. The package automatically converts R objects to 'NumPy' or 'PyTorch' tensors via 'reticulate' and supports GPU acceleration when 'PyTorch' with 'CUDA' is available. Python dependencies are declared via 'reticulate' and can be installed automatically on first use. Designed for large-scale statistical estimation where numerical stability and performance are critical.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cxy0714.github.io/U-Statistics-R/}
  \item \url{https://github.com/cxy0714/U-Statistics-R}
  \item \url{https://pypi.org/project/u-stats/}
  \item Report bugs at \url{https://github.com/cxy0714/U-Statistics-R/issues}
}

}
\author{
\strong{Maintainer}: Xingyu Chen \email{xingyuchen0714@sjtu.edu.cn}

Authors:
\itemize{
  \item Ruiqi Zhang \email{zrq1706@outlook.com}
  \item Lin Liu \email{linliu@sjtu.edu.cn}
}

}
\keyword{internal}
