% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\docType{data}
\name{ustats_env}
\alias{ustats_env}
\title{Internal package environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
ustats_env
}
\description{
A dedicated environment for storing ustats package state,
such as Python availability checks. Parent is emptyenv() for isolation.
}
\keyword{internal}
